+++ src/Opulence/Applications/Application.php
@@ -18,7 +18,7 @@ use Opulence\Applications\Tasks\TaskTypes;
 class Application
 {
     /** @var string The current Opulence version */
-    private static $opulenceVersion = "1.0.1";
+    private static $opulenceVersion = "1.0.2";
     /** @var ITaskDispatcher The task dispatcher */
     private $taskDispatcher = null;
     /** @var string The version of the application */