+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+<h2>v1.0.2 (2016-12-?)</h2>
+
+<h3>Console</h3>
+* Fixed bug that occurred with `null` short option names
+
 <h2>v1.0.1 (2016-12-11)</h2>
 
 <h3>HTTP</h3>